move download wasm file button to top part so it's visible for all tabs
